Improved changes
